// server/src/enums/areas.ts
export enum areas {
    EXATAS = "Ciências Exatas e da Terra",
    BIOLOGICAS = "Ciências Biológicas",
    ENGENHARIAS = "Engenharias",
    SAUDE = "Ciências da Saúde",
    AGRARIAS = "Ciências Agrárias",
    LETRAS = "Linguística, Letras e Artes",
    SOCIAIS = "Ciências Sociais Aplicadas",
    HUMANAS = "Ciências Humanas"
}